requirements.txt
langchain==0.3.14
openai==1.57.4
tiktoken==0.8.0
langchain_community==0.3.14

haystack-ai==2.8.0
streamlit==1.41.1
sentence-transformers>=3.0.0




